refactor the structure of the project, update 'requirements.txt'
