+++ requirements.txt
@@ -16,4 +16,4 @@ scipy==1.7.*
 seaborn==0.11.*
 tensorflow==2.8.*
 tqdm==4.64.*
-protobuf==3.20.*
+protobuf==3.19.*
