+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = syngen
-version = file: src/VERSION
+version = file: src/syngen/VERSION
 description = file: DESCRIPTION
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -43,11 +43,13 @@ install_requires =
     seaborn==0.11.2
     tensorflow==2.8.0
     tqdm==4.64.0
-    protobuf==3.20.*
+    protobuf==3.19.*
 [options.packages.find]
 where = src
+[options.package_data]
+* = *.py
 
 [options.entry_points]
 console_scripts =
-    train = train:train_model
-    infer = infer:infer_model
+    train = syngen.train:train_model
+    infer = syngen.infer:infer_model
