--- src/ml/config/__init__.py
@@ -1,2 +1,0 @@
-from ml.config.configurations import TrainConfig
-from ml.config.configurations import InferConfig
