--- src/ml/data_loaders/__init__.py
@@ -1,5 +1,0 @@
-from ml.data_loaders.data_loaders import CSVLoader
-from ml.data_loaders.data_loaders import AvroLoader
-from ml.data_loaders.data_loaders import JSONLoader
-from ml.data_loaders.data_loaders import DataLoader
-from ml.data_loaders.data_loaders import MetadataLoader
