--- src/ml/metrics/__init__.py
@@ -1,8 +1,0 @@
-from ml.metrics.metrics_classes.metrics import BaseMetric
-from ml.metrics.metrics_classes.metrics import BivariateMetric
-from ml.metrics.metrics_classes.metrics import UnivariateMetric
-from ml.metrics.metrics_classes.metrics import JensenShannonDistance
-from ml.metrics.metrics_classes.metrics import Correlations
-from ml.metrics.accuracy_test.accuracy_test import BaseTest
-from ml.metrics.accuracy_test.accuracy_test import AccuracyTest
-from ml.metrics.utils import encode_categorical_features
