--- src/ml/strategies/__init__.py
@@ -1,2 +1,0 @@
-from ml.strategies.strategies import TrainStrategy
-from ml.strategies.strategies import InferStrategy
