--- src/ml/train_chain/__init__.py
@@ -1,4 +1,0 @@
-from ml.train_chain.train_chain import BaseHandler
-from ml.train_chain.train_chain import RootHandler
-from ml.train_chain.train_chain import VaeTrainHandler
-from ml.train_chain.train_chain import VaeInferHandler
