--- src/ml/vae/models/__init__.py
@@ -1,1 +1,0 @@
-from ml.vae.models.dataset import Dataset
