--- src/ml/vae/wrappers/__init__.py
@@ -1,2 +1,0 @@
-from ml.vae.wrappers.wrappers import BaseWrapper
-from ml.vae.wrappers.wrappers import VAEWrapper
