+++ src/syngen/VERSION
@@ -1,0 +1,1 @@
+0.0.22
