+++ src/syngen/ml/config/__init__.py
@@ -1,0 +1,2 @@
+from syngen.ml.config.configurations import TrainConfig
+from syngen.ml.config.configurations import InferConfig
