+++ src/syngen/ml/metrics/__init__.py
@@ -1,0 +1,8 @@
+from syngen.ml.metrics.metrics_classes.metrics import BaseMetric
+from syngen.ml.metrics.metrics_classes.metrics import BivariateMetric
+from syngen.ml.metrics.metrics_classes.metrics import UnivariateMetric
+from syngen.ml.metrics.metrics_classes.metrics import JensenShannonDistance
+from syngen.ml.metrics.metrics_classes.metrics import Correlations
+from syngen.ml.metrics.accuracy_test.accuracy_test import BaseTest
+from syngen.ml.metrics.accuracy_test.accuracy_test import AccuracyTest
+from syngen.ml.metrics.utils import encode_categorical_features
