+++ src/syngen/ml/strategies/__init__.py
@@ -1,0 +1,2 @@
+from syngen.ml.strategies.strategies import TrainStrategy
+from syngen.ml.strategies.strategies import InferStrategy
