+++ src/syngen/ml/vae/__init__.py
@@ -1,0 +1,3 @@
+from syngen.ml.vae.wrappers.wrappers import BaseWrapper
+from syngen.ml.vae.wrappers.wrappers import VAEWrapper
+from syngen.ml.vae.wrappers.wrappers import VanillaVAEWrapper
