+++ src/syngen/ml/vae/models/__init__.py
@@ -1,0 +1,1 @@
+from syngen.ml.vae.models.dataset import Dataset
