+++ src/syngen/ml/vae/wrappers/__init__.py
@@ -1,0 +1,2 @@
+from syngen.ml.vae.wrappers.wrappers import BaseWrapper
+from syngen.ml.vae.wrappers.wrappers import VAEWrapper
